Send complete data struct
Now the whole struct is sent from the Arduino to the PC. The struct is
first serialized on the Arduino and then unpacked on the PC.

The message is introduced with the arbitrarily chosen character '0' to
signal that the packet is incoming and sync the data stream.

The PC responds with a string that the Arduino simply bounces back,
which the PC prints.