Run serial library with asyncio
Main is now changed so that the serial listening is happening in
parallel with a dummy function that prints a message every second. This
is in preparation for when the websocket library is intruduced which
uses asyncio.

I eventually ended up using the run_in_executor function to run the
blocking serial related function asynchronously.