Send serial from websocket handler
Yet another restructure of main. With the previous structure the serial
and websocket logic was completely separated. Now the websocket logic is
integrated into the serial logic so that they can interoperate.

A timeout is used when receiving websocket data so that it only spends
on receiving if there is anything to immediately receive.

The serial receive was rewritten to no use loops so as to not block the
asyncio event loop.