Split the main loop in two
The asyncronous nature of the code was basically removed with the
previous versions. Also, the websocket handle never exited when
the connection was lost so the incoming serial messages were read by two
processes and nothign worked.

Now when the websocket connects it spawns an async loop for the serial
connection. Now there are one loop for each connection and they are
properly exited.