Send start/stop command
The start/stop button now sends the command all the way to the CanSat.
No other commands are implemented yet.

Also moved the Arduino code for generating placeholder data into a new
folder and changed the order of data in the datastructure to be the same
as on the CanSat.