Change radio channel from GUI
There is now a select element where the dropdown contains all of our
allowed channels. When a new channel is picked, a radio channel command
is sent to the CanSat.

I've also made sure that the dropdown is disabled while the CanSat is
actively sending data. To change the channel the CanSat must first be
stopped.