Handle missing or strange data better
I've changed on the CanSat so that the data that is not ready is marked
with the same error state as when there is somethign wrong with the
sensor. This cange is necessary because the ground station can't know
what data to listen to from a sensor that is turned on.

I've also fixed so that values that are None are not saved to a .csv.

The last fix was to ignore data with no-negative timestamps. I
noticed that after subtracting the starting time some timestamps became
negative. I'll just remove them for now.