Fix temp/hum not showing in GUI
There were two issues with showing temperature and humidity in the GUI.

First, the data was only sending twice per second, and therefore missed
when the temperature or humidity changed. Now the GUI get new data
twice per second or when there is new temperature or humidity date.

Second, when there are gaps in the data, those gaps were not connected
with a line. Now all gaps in the data are connected with lines.